+++ testhelpers/observability/local_observability_endpoint.go
@@ -6,11 +6,11 @@ import (
 	"sync"
 
 	"github.com/google/uuid"
-	"github.com/grafana/oats/internal/testhelpers/common"
-	"github.com/grafana/oats/internal/testhelpers/otelcollector"
-	"github.com/grafana/oats/internal/testhelpers/prometheus"
-	"github.com/grafana/oats/internal/testhelpers/tempo"
 	"github.com/grafana/oats/observability"
+	"github.com/grafana/oats/testhelpers/common"
+	"github.com/grafana/oats/testhelpers/otelcollector"
+	"github.com/grafana/oats/testhelpers/prometheus"
+	"github.com/grafana/oats/testhelpers/tempo"
 	"go.opentelemetry.io/otel/sdk/resource"
 	"go.opentelemetry.io/otel/sdk/trace"
 )
