+++ testhelpers/otelcollector/otel_collector_config_template.go
