+++ testhelpers/otelcollector/otelcol-builder-manifest.yaml
