+++ testhelpers/tempo/local_tempo_endpoint.go
@@ -13,7 +13,7 @@ import (
 	"sync"
 	"text/template"
 
-	"github.com/grafana/oats/internal/testhelpers/common"
+	"github.com/grafana/oats/testhelpers/common"
 	"github.com/ory/dockertest/v3"
 	"github.com/ory/dockertest/v3/docker"
 	"go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracegrpc"
