+++ testhelpers/tempo/local_tempo_endpoint_test.go
@@ -1,28 +1,28 @@
-package otelcollector_test
+package tempo_test
 
 import (
 	"context"
 	"fmt"
+	"io"
+	"net/http"
 	"time"
 
 	"github.com/google/uuid"
-	"github.com/grafana/oats/internal/testhelpers/common"
-	"github.com/grafana/oats/internal/testhelpers/prometheus"
-	"github.com/grafana/oats/internal/testhelpers/tempo"
+	"github.com/grafana/oats/testhelpers/common"
+	"github.com/grafana/oats/testhelpers/prometheus"
 	"go.opentelemetry.io/otel/sdk/resource"
-	"google.golang.org/grpc"
 
 	semconv "go.opentelemetry.io/otel/semconv/v1.20.0"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
-	"github.com/grafana/oats/internal/testhelpers/otelcollector"
+	"github.com/grafana/oats/testhelpers/tempo"
 )
 
-var _ = Describe("provisioning an OpenTelemetry Collector locally using Docker", Label("integration", "docker", "slow"), func() {
+var _ = Describe("provisioning Tempo locally using Docker", Label("integration", "docker", "slow"), func() {
 	Describe("LocalEndpoint", func() {
-		It("can start a local OpenTelemetry Collector endpoint", func() {
+		It("can start a local Tempo endpoint", func() {
 			ctx := context.Background()
 
 			sharedNetworkUUID, err := uuid.NewRandom()
@@ -38,38 +38,33 @@ var _ = Describe("provisioning an OpenTelemetry Collector locally using Docker",
 			promEndpoint, err := prometheus.NewLocalEndpoint(ctx, sharedNetworkName)
 			Expect(err).ToNot(HaveOccurred(), "expected no error creating a local Prometheus endpoint")
 
-			promEndpointAddress, err := promEndpoint.Start(ctx)
+			promAddress, err := promEndpoint.Start(ctx)
 			Expect(err).ToNot(HaveOccurred(), "expected no error starting the local Prometheus endpoint")
 
 			defer promEndpoint.Stop(ctx)
 
-			tempoEndpoint, err := tempo.NewLocalEndpoint(ctx, sharedNetworkName, promEndpointAddress)
+			endpoint, err := tempo.NewLocalEndpoint(ctx, sharedNetworkName, promAddress)
 			Expect(err).ToNot(HaveOccurred(), "expected no error creating a local Tempo endpoint")
 
-			_, err = tempoEndpoint.Start(ctx)
+			endpointURL, err := endpoint.Start(ctx)
 			Expect(err).ToNot(HaveOccurred(), "expected no error starting a local Tempo endpoint")
 
-			defer tempoEndpoint.Stop(ctx)
+			defer endpoint.Stop(ctx)
 
-			traceEndpoint, err := tempoEndpoint.OTLPTraceEndpoint(ctx)
-			Expect(err).ToNot(HaveOccurred(), "expected no error getting the endpoint for sending traces to Tempo")
+			resp, err := http.Get(fmt.Sprintf("http://%s/status", endpointURL.HostEndpoint))
+			Expect(err).ToNot(HaveOccurred(), "expected no error getting the status of the local Tempo endpoint")
 
-			collectorEndpoint, err := otelcollector.NewLocalEndpoint(ctx, sharedNetworkName, traceEndpoint, promEndpointAddress)
-			Expect(err).ToNot(HaveOccurred(), "expeccted no error creating a local OpenTelemetry Collector endpoint")
+			defer resp.Body.Close()
 
-			collectorAddress, err := collectorEndpoint.Start(ctx)
-			Expect(err).ToNot(HaveOccurred(), "expected no error starting a local OpenTelemetry Collector endpoint")
+			Expect(resp.StatusCode).To(Equal(http.StatusOK), "expected 200 OX from the local Tempo endpoint")
 
-			defer collectorEndpoint.Stop(ctx)
+			respBytes, err := io.ReadAll(resp.Body)
+			Expect(err).ToNot(HaveOccurred(), "expected no error reading a Tempo status response")
 
-			conn, err := grpc.Dial(collectorAddress.HostEndpoint, grpc.WithInsecure())
-			Expect(err).ToNot(HaveOccurred(), "expected no error connecting to the gRPC endpoint of the OpenTelemetry Collector")
-
-			err = conn.Close()
-			Expect(err).ToNot(HaveOccurred(), "expected no error closing the gRPC connection to the OpenTelemetry Collector")
+			Expect(string(respBytes)).To(ContainSubstring("tempo, version 2.1.1"), "expected to get the Tempo version from the status endpoint")
 		})
 
-		It("provides an OpenTelemetry TracerProvider for sending trace data", func() {
+		It("provides an OpenTelemetry TraceProvider for sending trace data, and an endpoint for returning a trace by ID", func() {
 			ctx := context.Background()
 
 			sharedNetworkUUID, err := uuid.NewRandom()
@@ -85,54 +80,43 @@ var _ = Describe("provisioning an OpenTelemetry Collector locally using Docker",
 			promEndpoint, err := prometheus.NewLocalEndpoint(ctx, sharedNetworkName)
 			Expect(err).ToNot(HaveOccurred(), "expected no error creating a local Prometheus endpoint")
 
-			promEndpointAddress, err := promEndpoint.Start(ctx)
+			promAddress, err := promEndpoint.Start(ctx)
 			Expect(err).ToNot(HaveOccurred(), "expected no error starting the local Prometheus endpoint")
 
 			defer promEndpoint.Stop(ctx)
 
-			tempoEndpoint, err := tempo.NewLocalEndpoint(ctx, sharedNetworkName, promEndpointAddress)
+			endpoint, err := tempo.NewLocalEndpoint(ctx, sharedNetworkName, promAddress)
 			Expect(err).ToNot(HaveOccurred(), "expected no error creating a local Tempo endpoint")
 
-			_, err = tempoEndpoint.Start(ctx)
+			_, err = endpoint.Start(ctx)
 			Expect(err).ToNot(HaveOccurred(), "expected no error starting a local Tempo endpoint")
 
-			defer tempoEndpoint.Stop(ctx)
-
-			traceEndpoint, err := tempoEndpoint.OTLPTraceEndpoint(ctx)
-			Expect(err).ToNot(HaveOccurred(), "expected no error getting the endpoint for sending traces to Tempo")
-
-			collectorEndpoint, err := otelcollector.NewLocalEndpoint(ctx, sharedNetworkName, traceEndpoint, promEndpointAddress)
-			Expect(err).ToNot(HaveOccurred(), "expeccted no error creating a local OpenTelemetry Collector endpoint")
-
-			_, err = collectorEndpoint.Start(ctx)
-			Expect(err).ToNot(HaveOccurred(), "expected no error starting a local OpenTelemetry Collector endpoint")
-
-			defer collectorEndpoint.Stop(ctx)
+			defer endpoint.Stop(ctx)
 
 			r, err := resource.Merge(
 				resource.Default(),
 				resource.NewWithAttributes(
 					"", // use the SchemaURL from the default resource
-					semconv.ServiceName("LocalOTelCollectorEndpointTest"),
+					semconv.ServiceName("LocalTempoEndpointTest"),
 				),
 			)
 
 			Expect(err).ToNot(HaveOccurred(), "expected no error creating an OpenTelemetry resource")
 
-			traceProvider, err := collectorEndpoint.TracerProvider(ctx, r)
+			traceProvider, err := endpoint.TracerProvider(ctx, r)
 			Expect(err).ToNot(HaveOccurred(), "expected no error getting a trace provider")
 
 			defer traceProvider.Shutdown(ctx)
 
-			tracer := traceProvider.Tracer("LocalOTelCollectorEndpointTestTracer")
+			tracer := traceProvider.Tracer("LocalTempoEndpointTestTracer")
 
-			parentCtx, parentSpan := tracer.Start(ctx, "local-otel-collector-endpoint-test-parent")
+			parentCtx, parentSpan := tracer.Start(ctx, "local-tempo-endpoint-test-parent")
 
-			const eventMessage = "taking a little siesta"
+			const eventMessage = "taking a little nap"
 
 			// create a closure over the tracer, parent context, and event message
-			helloOtel := func() {
-				_, childSpan := tracer.Start(parentCtx, "hello-otel")
+			helloTempo := func() {
+				_, childSpan := tracer.Start(parentCtx, "hello-tempo")
 				defer childSpan.End()
 
 				childSpan.AddEvent(eventMessage)
@@ -140,27 +124,26 @@ var _ = Describe("provisioning an OpenTelemetry Collector locally using Docker",
 				time.Sleep(250 * time.Millisecond)
 			}
 
-			helloOtel()
+			helloTempo()
 			parentSpan.End()
 
 			parentSpanContext := parentSpan.SpanContext()
-			Expect(parentSpanContext.HasTraceID()).To(BeTrue(), "expected the parent local OpenTelemetry Collector endpoint test span to have a valid TraceID")
+			Expect(parentSpanContext.HasTraceID()).To(BeTrue(), "expected the parent local tempo endpoint test span to have a valid TraceID")
 
 			err = traceProvider.ForceFlush(ctx)
 			Expect(err).ToNot(HaveOccurred(), "expected no error flushing the trace provider")
 
 			traceID := parentSpanContext.TraceID()
 
-			fetchedTrace, err := tempoEndpoint.GetTraceByID(ctx, traceID.String())
-			if err != nil {
-				fmt.Println("there was an error fetching the trace from Tempo, you have 60s to investigate.....")
-				time.Sleep(60 * time.Second)
-			}
-
+			fetchedTrace, err := endpoint.GetTraceByID(ctx, traceID.String())
 			Expect(err).ToNot(HaveOccurred(), "expected no error fetching the exported trace from Tempo")
 
 			Expect(string(fetchedTrace)).ToNot(BeEmpty(), "expected a non empty response from Tempo when getting an exported trace by ID")
 			Expect(string(fetchedTrace)).To(ContainSubstring(eventMessage), "expected the event message to be contained in the returned trace")
+		})
+
+		XIt("returns the gRPC endpoint for sending traces to", func() {
+			Fail("test not written")
 		})
 
 		XIt("can be stopped", func() {
@@ -168,6 +151,10 @@ var _ = Describe("provisioning an OpenTelemetry Collector locally using Docker",
 		})
 
 		XIt("tries to respect context cancellation", func() {
+			Fail("test not written")
+		})
+
+		XIt("starts idempotently", func() {
 			Fail("test not written")
 		})
 	})
